feat: simplify TriggerePublisher onComplete
