+++ executions/graphql-kotlin-transaction-batcher/src/main/kotlin/com/expediagroup/graphql/transactionbatcher/transaction/cache/DefaultTransactionBatcherCache.kt
@@ -14,14 +14,22 @@
  * limitations under the License.
  */
 
-package com.expediagroup.graphql.transactionbatcher.transaction
+package com.expediagroup.graphql.transactionbatcher.transaction.cache
+
+import java.util.concurrent.ConcurrentHashMap
 
 /**
- * Interface that allows any cache implementation that can be used by the [TransactionBatcher]
- * by default will use [DefaultTransactionBatcherCache]
+ * Default implementation of [TransactionBatcherCache] using an in memory [cache]
+ * without eviction
  */
-interface TransactionBatcherCache {
-    fun set(key: String, value: Any)
-    fun get(key: String): Any?
-    fun contains(key: String): Boolean
+class DefaultTransactionBatcherCache : TransactionBatcherCache {
+    private val cache = ConcurrentHashMap<String, Any>()
+
+    override fun set(key: String, value: Any) {
+        cache[key] = value
+    }
+
+    override fun get(key: String): Any? = cache[key]
+
+    override fun contains(key: String): Boolean = cache.containsKey(key)
 }
